% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cvs.R
\name{calculate_moes}
\alias{calculate_moes}
\title{Calculate margins of error for derived variables}
\usage{
calculate_moes(.df)
}
\arguments{
\item{.df}{The dataset returned from \code{compile_acs_data()}.
The argument to this parameter must have an attribute named \code{codebook} (as is
true of results from \code{compile_acs_data())}.}
}
\value{
A modified dataframe that includes margins of error (suffixed
\code{_M}) for derived variables.
}
\description{
Calculate margins of error for derived variables
}
\details{
Calculates margins of error for all derived ACS estimates. Standard
errors are computed internally as an intermediate step but are not included
in the returned dataframe. Uses pre-parsed codebook columns
(numerator_vars, denominator_vars, se_calculation_type) to determine how
to calculate standard errors.
}
\keyword{internal}
