% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_registry.R
\name{define_complement}
\alias{define_complement}
\title{Define a complement (1 - x) variable}
\usage{
define_complement(source, output)
}
\arguments{
\item{source}{A string. The column to subtract from 1.}

\item{output}{A string. The name of the output column to create.}
}
\value{
A list with \code{type = "one_minus"} and the associated fields.
Can be passed in the \code{tables} parameter of
\code{\link{compile_acs_data}}.
}
\description{
Creates a definition that computes \code{1 - source}.
}
