% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_registry.R
\name{define_one_minus}
\alias{define_one_minus}
\title{Define a one-minus (complement) variable}
\usage{
define_one_minus(output, source_variable)
}
\arguments{
\item{output}{A string. The name of the output column to create.}

\item{source_variable}{A string. The column to subtract from 1.}
}
\value{
A list with \code{type = "one_minus"} and the associated fields.
Can be passed in the \code{tables} parameter of
\code{\link{compile_acs_data}}.
}
\description{
Creates a definition that computes \code{1 - source_variable}.
}
