% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_registry.R
\name{define_percent}
\alias{define_percent}
\title{Define a percentage variable (simple or complex)}
\usage{
define_percent(
  output,
  numerator = NULL,
  denominator = NULL,
  numerator_variables = NULL,
  numerator_regex = NULL,
  numerator_exclude_regex = NULL,
  numerator_note = NULL,
  numerator_subtract_variables = NULL,
  numerator_subtract_regex = NULL,
  denominator_variables = NULL,
  denominator_regex = NULL,
  denominator_exclude_regex = NULL,
  subtract_variables = NULL,
  subtract_regex = NULL
)
}
\description{
Define a percentage variable (simple or complex)
}
\keyword{internal}
