% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_acs_variables.R
\name{filter_variables}
\alias{filter_variables}
\title{Easily filter ACS variables}
\usage{
filter_variables(variable_vector, match_string, match_type = "positive")
}
\arguments{
\item{variable_vector}{A named vector (intended for use with named ACS variables).}

\item{match_string}{A string on which to filter (or not filter) elements in \code{variable_vector}.}

\item{match_type}{Whether to include (\code{match_type = "positive"}) or exclude
(\code{match_type = "negative"}) matching elements.}
}
\value{
The elements from \code{variable_vector} that do/don't match \code{match_string}.
}
\description{
Filter the the results of \code{select_variables_by_name()} based on their \code{match_type} relative
to \code{match_string}.
}
\examples{
\dontrun{
codebook = tidycensus::load_variables(dataset = "acs5", year = 2022)
selected_variables = select_variables_by_name("B16005_", census_codebook = codebook)
filter_variables(
  variable_vector = selected_variables,
  match_string = "universe_$|native_$|foreign_born_$|only|very_well",
  match_type = "positive")
}
}
