% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_acs_data.R
\name{internal_compute_acs_variables}
\alias{internal_compute_acs_variables}
\title{Calculate ACS measures}
\usage{
internal_compute_acs_variables(.data)
}
\arguments{
\item{.data}{The dataset returned from `compile_acs_data()`.}
}
\value{
A modified dataframe that includes newly calculated indicators.
}
\description{
Calculates derived ACS indicators.
}
\details{
An internal function used to calculate indicators and derived variable
definitions for the codebook.
}
\examples{
\dontrun{
df = compile_acs_data(
  variables = list_acs_variables(year = 2022),
  years = c(2022),
  geography = "county",
  states = "NJ",
  counties = NULL,
  retain_moes = TRUE,
  spatial = FALSE)
internal_compute_acs_variables(.data = df)
}
}
