% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pretty_names.R
\name{make_pretty_names}
\alias{make_pretty_names}
\title{Pretty-ify variable names}
\usage{
make_pretty_names(.data, .case = "title")
}
\arguments{
\item{.data}{A data.frame--or something coercible thereto--or a character vector containing
variables/variable names that will be converted more publication-appropriate formats.}

\item{.case}{Capitalization scheme of resulting variable names. One of "title", "sentence", or "upper".}
}
\description{
Variable names are lower-case and underscore-separated. This converts
them to title case and replaces underscores with spaces. It also makes other adjustments
to facilitate reading and interpretation of variables in the context of, for example,
plots and tables.
}
\examples{
\dontrun{
"race_personofcolor_percent" \%>\% make_pretty_names()
}
}
