% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cvs.R
\name{se_product}
\alias{se_product}
\title{Calculate standard error for a product of two estimates}
\usage{
se_product(
  estimate_x,
  estimate_y,
  se_x = NULL,
  se_y = NULL,
  moe_x = NULL,
  moe_y = NULL
)
}
\arguments{
\item{estimate_x}{The first estimate (X)}

\item{estimate_y}{The second estimate (Y)}

\item{se_x}{The standard error of estimate X (or NULL if providing MOE)}

\item{se_y}{The standard error of estimate Y (or NULL if providing MOE)}

\item{moe_x}{The margin of error of estimate X (or NULL if providing SE)}

\item{moe_y}{The margin of error of estimate Y (or NULL if providing SE)}
}
\value{
The standard error of the product X*Y
}
\description{
Calculate standard error for a product of two estimates
}
\details{
Calculate the standard error for an estimate derived by multiplying
two estimates together. For example, multiplying a proportion by a population
count to get a subgroup count. Formula from Census Bureau ACS Accuracy
documentation: SE(X\emph{Y) = sqrt((X}SE(Y))^2 + (Y*SE(X))^2).
}
\keyword{internal}
