% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cvs.R
\name{se_simple}
\alias{se_simple}
\title{Calculate a simple standard error}
\usage{
se_simple(moe)
}
\arguments{
\item{moe}{A margin of error, or a vector thereof}
}
\value{
A 90\% standard error
}
\description{
Calculate a simple standard error
}
\details{
Create a standard error at the 90\% level from a 90\% margin of error
}
\keyword{internal}
