% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_acs_variables.R
\name{select_variables_by_name}
\alias{select_variables_by_name}
\title{Given the name of an ACS variable (or a string that matches one or more such variables),
generate a named character vector of original variable names and more meaningful names.}
\usage{
select_variables_by_name(variable_name, census_codebook)
}
\arguments{
\item{variable_name}{A named vector (intended for use with named ACS variables).}

\item{census_codebook}{An object returned from [tidycensus::load_variables()].}
}
\value{
A named character vector containing the variables that matched
   `variable_name` from the `census_codebook`, with semantically-meaningful names
   derived from metadata fields contained in `census_codebook`.
}
\description{
Given the name of an ACS variable (or a string that matches one or more such variables),
generate a named character vector of original variable names and more meaningful names.
}
\examples{
codebook = tidycensus::load_variables(dataset = "acs5", year = 2022)
select_variables_by_name("B16005_", census_codebook = codebook)
}
