% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cvs.R
\name{calculate_cvs}
\alias{calculate_cvs}
\title{Calculate Coefficients of Variation (CVs)}
\usage{
calculate_cvs(.df)
}
\arguments{
\item{.df}{The dataset returned from \code{compile_acs_data()}.
The argument to this parameter must have an attribute named \code{codebook} (as is)
true of results from \code{compile_acs_data()}.}
}
\value{
A modified dataframe that includes newly calculated indicators.
}
\description{
Calculate Coefficients of Variation (CVs)
}
\details{
Create CVs for all ACS estimates and derived indicators
}
\examples{
\dontrun{
df = compile_acs_data(
  variables = list_acs_variables(year = 2022),
  years = c(2022),
  geography = "county",
  states = "NJ",
  counties = NULL,
  spatial = FALSE)
cvs = calculate_cvs(df) \%>\%
 dplyr::select(matches("_cv$"))
}
}
