% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_acs_data.R
\name{compile_acs_data}
\alias{compile_acs_data}
\title{Analysis-ready social science measures}
\usage{
compile_acs_data(
  variables = NULL,
  years = c(2022),
  geography = "county",
  states = NULL,
  counties = NULL,
  retain_moes = TRUE
)
}
\arguments{
\item{variables}{A named vector of ACS variables such as that returned from
[urbnindicators::list_acs_variables()].}

\item{years}{A character vector (or coercible to the same) comprising one or more
four-digit years for which to pull five-year American Community Survey estimates.}

\item{geography}{A geography type that is accepted by [tidycensus::get_acs()], e.g.,
"tract", "county", "state", among others.}

\item{states}{A vector of one or more state names, abbreviations, or codes as
accepted by [tidycensus::get_acs()].}

\item{counties}{A vector of one or more county names, abbreviations, or codes as
accepted by [tidycensus::get_acs()]. NOTE: this parameter is not currently supported
and must be set to NULL (as it is by default).}

\item{retain_moes}{Boolean. Include margins of error (MOE) in the returned dataframe,
or omit them?}
}
\value{
A dataframe containing the requested `variables`, their MOEs (optionally),
   a series of derived variables, such as percentages, and the year of the data.
   Returned data are formatted wide. A codebook generated with `generate_codebook()`
   is attached and can be accessed via `compile_acs_data() %>% attr("codebook")`.
}
\description{
`compile_acs_data()` constructs measures frequently used in social sciences
   research while leveraging [tidycensus::get_acs()] to acquire raw estimates from
   the Census Bureau API.
}
\examples{
acs_variables = list_acs_variables(year = "2022")
df = compile_acs_data(
  variables = acs_variables,
  years = c(2021, 2022),
  geography = "county",
  states = "NJ",
  counties = NULL,
  retain_moes = TRUE)
codebook = df \%>\%
  attr("codebook")
}
\seealso{
[tidycensus::get_acs()], which this function wraps.
}
