% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cvs.R
\name{cv}
\alias{cv}
\title{Calculate a coefficient of variation}
\usage{
cv(estimate, se)
}
\arguments{
\item{estimate}{The estimate}

\item{se}{The standard error}
}
\value{
A coefficient of variation at the 90\% level
}
\description{
Calculate a coefficient of variation
}
\details{
Return a coefficient of variation at the 90\% level
}
