% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_registry.R
\name{define_across_percent}
\alias{define_across_percent}
\title{Define an across-percent variable}
\usage{
define_across_percent(
  input_regex,
  output_suffix,
  denominator = NULL,
  denominator_function = NULL,
  denominator_subtract = NULL,
  exclude_regex = NULL
)
}
\description{
Define an across-percent variable
}
\keyword{internal}
