% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_registry.R
\name{define_across_sum}
\alias{define_across_sum}
\title{Define an across-sum variable}
\usage{
define_across_sum(
  input_regex,
  addend_function,
  output_naming_function,
  exclude_regex = NULL
)
}
\description{
Define an across-sum variable
}
\keyword{internal}
