% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_acs_variables.R
\name{filter_variables}
\alias{filter_variables}
\title{Filter the the results of [select_variables()] based on their `match_type` relative
to `match_string`.}
\usage{
filter_variables(variable_vector, match_string, match_type = "positive")
}
\arguments{
\item{variable_vector}{A named vector (intended for use with named ACS variables).}

\item{match_string}{a string on which to filter (or not filter) elements in `variable_vector`.}

\item{match_type}{whether to include (`match_type = "positive"`) or exclude
(`match_type = "negative"`) matching elements}
}
\value{
The elements from `variable_vector` that do/don't match `match_string`.
}
\description{
Filter the the results of [select_variables()] based on their `match_type` relative
to `match_string`.
}
\examples{
codebook = tidycensus::load_variables(dataset = "acs5", year = 2022)
variables = select_variables_by_name("B16005_", census_codebook = codebook)
filter_variables(
  variable_vector = variables,
  match_string = "universe_$|native_$|foreign_born_$|only|very_well",
  match_type = "positive")
}
