% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_codebook.R
\name{generate_codebook}
\alias{generate_codebook}
\title{Document variables from `compile_acs_data()`}
\usage{
generate_codebook(.data)
}
\arguments{
\item{.data}{The dataset returned from `compile_acs_data()`.}
}
\value{
A tibble containing the names and definitions of  variables returned from
`compile_acs_data()`.
}
\description{
`generate_codebook()` defines how variables produced via `compile_acs_data()`
are calculated.
}
\details{
Generates a tibble of variable names and definitions that describe
how each variable was created.
}
\examples{
\dontrun{
df = compile_acs_data(
  variables = list_acs_variables(year = 2022),
  years = c(2022),
  geography = "county",
  states = "NJ",
  counties = NULL,
  retain_moes = TRUE,
  spatial = FALSE)
codebook = generate_codebook(.data = df)
}
}
