% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_codebook.R
\name{generate_codebook}
\alias{generate_codebook}
\title{Document variables from \code{urbnindicators::compile_acs_data()}}
\usage{
generate_codebook(.data)
}
\arguments{
\item{.data}{The dataset returned from \code{urbnindicators::compile_acs_data()}.}
}
\value{
A tibble containing the names and definitions of  variables returned from
\code{urbnindicators::compile_acs_data()}.
}
\description{
Define how variables produced via \code{urbnindicators::compile_acs_data()}
are calculated.
}
\details{
Generates a tibble of variable names and definitions that describe
how each variable was created.
}
\examples{
\dontrun{
df = compile_acs_data(
  variables = list_acs_variables(year = 2022),
  years = c(2022),
  geography = "county",
  states = "NJ",
  counties = NULL,
  spatial = FALSE)
codebook = generate_codebook(.data = df)
}
}
