% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_acs_variables.R
\name{list_acs_variables}
\alias{list_acs_variables}
\title{Return ACS variables codes and names}
\usage{
list_acs_variables(year = "2022")
}
\arguments{
\item{year}{The year for which variable names should be selected.}
}
\value{
A named vector of variable codes (as specified in the Census Bureau's API)
   with semantically-meaningful names (e.g., "race_black_alone_nonhispanic").
}
\description{
Generate meaningful names for ACS variable codes based on their metadata
   and return these as a vector, along with their semantic names. Intended for
   use with [compile_acs_data()].
}
\examples{
list_acs_variables(year = "2022")
}
