% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_acs_data.R
\name{safe_divide}
\alias{safe_divide}
\title{Division without NAs}
\usage{
safe_divide(x, y)
}
\arguments{
\item{x}{A numeric scalar.}

\item{y}{A numeric scalar.}
}
\value{
The traditional dividend in all cases except where \code{y == 0}, in which
case it returns 0.
}
\description{
Return 0 when the divisor is 0.
}
\details{
A modified division operation that returns zero when the divisor is zero
rather than returning NA. Otherwise returns the quotient.
}
\examples{
safe_divide(1, 2)
safe_divide(3, 0)
}
