% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cvs.R
\name{se_proportion_ratio}
\alias{se_proportion_ratio}
\title{Calculate a pooled standard error for a proportion or ratio}
\usage{
se_proportion_ratio(
  estimate_numerator,
  estimate_denominator,
  moe_numerator = NULL,
  moe_denominator = NULL,
  se_numerator = NULL,
  se_denominator = NULL,
  type = "proportion"
)
}
\arguments{
\item{estimate_numerator}{The estimate of the numerator}

\item{estimate_denominator}{The estimate of the denominator}

\item{moe_numerator}{The margin of error of the numerator}

\item{moe_denominator}{The margin of error of the denominator}

\item{se_numerator}{The standard error of the numerator}

\item{se_denominator}{The standard error of the denominator}

\item{type}{The type of estimate being calculated, either "proportion" or "ratio"}
}
\value{
A pooled 90\% standard error
}
\description{
Calculate a pooled standard error for a proportion or ratio
}
\details{
For an estimate derived using division--whether the resulting estimate is a proportion or a ratio--calculate the pooled standard error. While there are convenience parameters that support both SEs and MOEs for the numerator and denominator, only one of these--either SEs or MOEs--can be supplied--the other must be left NULL.
}
