% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cvs.R
\name{se_sum}
\alias{se_sum}
\title{Calculate a pooled standard error for a summed or subtracted estimate}
\usage{
se_sum(...)
}
\arguments{
\item{...}{The unquoted names of the margin of error and corresponding estimate variables to be added or subtracted}
}
\value{
A pooled 90\% standard error
}
\description{
Calculate a pooled standard error for a summed or subtracted estimate
}
\details{
For an estimate derived by adding or subtracting multiple estimates, calculate the pooled standard error
}
