% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urbnindicators-package.R
\docType{package}
\name{urbnindicators-package}
\alias{urbnindicators}
\alias{urbnindicators-package}
\title{urbnindicators: Out of the Box Social Science Indicators from the American Community Survey (ACS)}
\description{
There are many packages available that facilitate queries to the Census Bureau API, but subsequent steps in the analysis workflow still require significant work (much of which must be repeated from analysis to analysis), such as identifying and selecting relevant variables, determining how to accurately calculate derived measures (e.g., percentages), and naming variables semantically. urbnindicators abstracts much of this process away and provides users a simple interface to obtain commonly-used social sciences measures from the ACS.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ui-research.github.io/urbnindicators/}
}

}
\author{
\strong{Maintainer}: Will Curran-Groome \email{wcurrangroome@gmail.com}

Authors:
\itemize{
  \item Will Curran-Groome \email{wcurrangroome@gmail.com}
}

}
\keyword{internal}
